+++ src/cider/nrepl/middleware/util/nrepl.clj
@@ -1,9 +1,18 @@
 (ns cider.nrepl.middleware.util.nrepl
-  "Common utilities for interaction with the client."
+  "Common nREPL-related utilities."
   (:require
    [nrepl.middleware.interruptible-eval :refer [*msg*]]
    [nrepl.misc :refer [response-for]]
-   [nrepl.transport :as transport]))
+   [nrepl.transport :as transport]
+   [nrepl.version :refer [version]]))
+
+(defn satisfies-version?
+  "Check if the nREPL version is of the provided major and minor parts or newer."
+  [major minor]
+  (>= (compare ((juxt :major :minor) version) [major minor]) 0))
+
+#_(satisfies-version? 0 9)
+#_(satisfies-version? 1 10)
 
 (defn notify-client
   "Send user level notification to client as a response to request `msg`.