+++ test/clj/cider/nrepl/middleware/inspect_test.clj
@@ -598,18 +598,19 @@
 
 (deftest inspect-print-current-value-test
   (testing "inspect-print-current-value returns the currently inspected value as a printed string"
-    (is (= [(str/join "\n" ["{:a -1,"
-                            " :bb \"111\","
-                            " :ccc (1),"
-                            " :d"
-                            " ({:a 0, :bb \"000\", :ccc ()}"
-                            "  {:a -1, :bb \"111\", :ccc (1)}"
-                            "  {:a -2, :bb \"222\", :ccc (2 1)}"
-                            "  {:a -3, :bb \"333\", :ccc (3 2 1)}"
-                            "  {:a -4, :bb \"444\", :ccc (4 3 2 1)})}"])]
-           (:value (do
-                     (session/message {:op "eval"
-                                       :code "(def test-val
+    (is+ {:value ["{:a -1,
+ :bb \"111\",
+ :ccc (1),
+ :d
+ ({:a 0, :bb \"000\", :ccc ()}
+  {:a -1, :bb \"111\", :ccc (1)}
+  {:a -2, :bb \"222\", :ccc (2 1)}
+  {:a -3, :bb \"333\", :ccc (3 2 1)}
+  {:a -4, :bb \"444\", :ccc (4 3 2 1)})}
+"]}
+         (do
+           (session/message {:op "eval"
+                             :code "(def test-val
                                                 (for [i (range 2)]
                                                  {:a (- i)
                                                   :bb (str i i i)
@@ -618,13 +619,14 @@
                                                        {:a (- i)
                                                         :bb (str i i i)
                                                         :ccc (range i 0 -1)})}))"})
-                     (session/message {:op "eval"
-                                       :inspect "true"
-                                       :code "test-val"})
-                     (session/message {:op "inspect-push"
-                                       :idx 2})
-                     (session/message {:op "inspect-print-current-value"
-                                       :nrepl.middleware.print/print "cider.nrepl.pprint/pprint"})))))))
+           (session/message {:op "eval"
+                             :inspect "true"
+                             :code "test-val"})
+           (session/message {:op "inspect-push"
+                             :idx 2})
+           (session/message {:op "inspect-print-current-value"
+                             :nrepl.middleware.print/buffer-size 2048
+                             :nrepl.middleware.print/print "cider.nrepl.pprint/orchard-pprint"})))))
 
 (deftest inspect-print-current-value-no-value-test
   (testing "inspect-print-current-value returns nil if nothing has been inspected yet"