+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Bump `orchard` to [0.37.0](https://github.com/clojure-emacs/orchard/blob/master/CHANGELOG.md#0370-2025-09-19).
+* Bump `compliment` to [0.7.1](https://github.com/alexander-yakushev/compliment/blob/master/CHANGELOG.md#071-2025-09-19).
 
 ## 0.57.0 (2025-06-29)